# servers/go-server-server/go/user.go
/*
 * Simple Storage Server
 *
 * A Simple Storage Server API
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type User struct {

	Username string `json:"username,omitempty"`

	Password string `json:"password,omitempty"`
}